+++ apps/lemon_agent/test/test_helper.exs
@@ -1,6 +1,3 @@
 # By default, exclude integration tests that require external CLIs/services.
 ExUnit.configure(exclude: [:integration])
 ExUnit.start()
-
-# Compile and load support files
-Code.require_file("support/mocks.ex", __DIR__)